Move Path utility methods
